Set up Spring IOC container
