+++ WebContent/WEB-INF/web.xml
@@ -15,6 +15,18 @@
 	<!-- <location>/index.html</location> -->
 	<!-- </error-page> -->
 
+	<context-param>
+		<param-name>contextConfigLocation</param-name>
+		<param-value>/WEB-INF/applicationContext*.xml</param-value>
+	</context-param>
+
+	<listener>
+		<listener-class>
+			org.springframework.web.context.ContextLoaderListener
+		</listener-class>
+	</listener>
+
+
 	<servlet>
 		<servlet-name>PoguesBackOffice</servlet-name>
 		<servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
@@ -82,9 +94,9 @@
 			<http-method>POST</http-method>
 		</web-resource-collection>
 
-		<auth-constraint>
-			<role-name>Utilisateurs_Pogues</role-name>
-		</auth-constraint>
+		<!--<auth-constraint>-->
+			<!--<role-name>Utilisateurs_Pogues</role-name>-->
+		<!--</auth-constraint>-->
 
 	</security-constraint>
